# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentVault API",
    "version": "0.2.0",
    "description": "HTTP API for AgentVault — encrypted AI agent memory on Story Protocol. All POST endpoints require CSRF-safe Origin/Referer headers (allow-list: localhost:3000, NEXT_PUBLIC_SITE_URL, VERCEL_URL). All endpoints that hit the chain are rate-limited per IP.\n\n## Architecture\n\nAgentVault exposes endpoints for:\n- **CDR** — Confidential Data Rails threshold encryption (store / recall memory on-chain)\n- **Story** — IP Asset + License Token registration (user-pays: returns calldata templates)\n- **LLM** — Anthropic Sonnet inference proxy\n- **Contract** — On-chain reads of the AgentVault.sol registry (getUserIpIds, getUserAgents, getIpInfo, getAgent)\n- **Wallet** — IP balance check only (no in-app drip; users fund via external faucet)\n- **License** — Enumerate license tokens held by a wallet\n- **Marketplace** — Pre-encode license minting calldata for buyers to sign\n\n## Cross-cutting headers\n\n| Header | Required | Notes |\n|--------|----------|-------|\n| `Content-Type` | yes (POST) | `application/json` |\n| `Origin` | yes (POST) | Must match CSRF allow-list |\n\n## Error format\n\nAll errors return `{ \"error\": \"<message>\" }` with an appropriate 4xx/5xx status. Rate-limited responses include `resetMs` (ms until limit resets) or `retryInMs` / `retryInMin` for wallet cooldowns.",
    "contact": {
      "name": "AgentVault",
...
    },
    "license": {
...
    }
  },
  "servers": [
...
  "paths": {
...
    "/api/story/setup": {
      "post": {
        "tags": ["Story"],
        "summary": "Mint NFT, register IP Asset, attach license terms, mint license token, create AgentVault agent, store memory reference",
        "description": "USER-PAYS spawn flow. Returns pre-encoded calldata templates (ABIs + arg templates) for the 5 user-signed txs: mint NFT, register IP, attachLicense (via IPAccount.execute), mintLicense (via IPAccount.execute), and createAgentAndStoreMemory on the AgentVault contract. The user fills in tokenId/ipId placeholders after each receipt and signs sequentially. The server does NOT sign ownership txs (no more 'deployer drained' issue). After all 5 txs are mined, the client calls /api/story/finalize to register the IP in the AgentVault on-chain registry (1 server-signed tx). Schema below is approximate — the actual response contains viem-compatible `abi` + `args` templates for each step. For exact field names, see src/app/api/story/setup/route.ts.",
        "operationId": "storySetup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StorySetupRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Story setup complete",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StorySetupResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/CSRF" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/contract": {
      "post": {
        "tags": ["Contract"],
        "summary": "Read on-chain data from the AgentVault contract",
        "description": "Dispatcher endpoint. Pass `action` to select which read to perform against the deployed AgentVault.sol. Available actions: `getUserAgents`, `checkAccess`, `getAgentMemoryCount`.",
        "operationId": "contractRead",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  { "$ref": "#/components/schemas/ContractGetUserAgentsRequest" },
                  { "$ref": "#/components/schemas/ContractCheckAccessRequest" },
                  { "$ref": "#/components/schemas/ContractGetAgentMemoryCountRequest" }
                ],
                "discriminator": {
                  "propertyName": "action",
                  "mapping": {
                    "getUserAgents": "#/components/schemas/ContractGetUserAgentsRequest",
                    "checkAccess": "#/components/schemas/ContractCheckAccessRequest",
                    "getAgentMemoryCount": "#/components/schemas/ContractGetAgentMemoryCountRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result depends on action",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContractResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/CSRF" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/wallet/balance": {
      "post": {
        "tags": ["Wallet"],
        "summary": "Check native IP balance of a wallet",
        "description": "Returns the IP balance and whether it meets the minimum required to spawn an agent (currently 0.05 IP).",
        "operationId": "walletBalance",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WalletAddressRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Balance",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WalletBalanceResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/CSRF" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "string",
        "description": "0x-prefixed EVM address (20 bytes hex)",
        "pattern": "^0x[a-fA-F0-9]{40}$",
        "example": "0x936CDB5dD5DCE69a2DeC06299C986E7798ab274B"
      },
      "TxHash": {
        "type": "string",
        "description": "0x-prefixed transaction hash (32 bytes hex)",
        "pattern": "^0x[a-fA-F0-9]{64}$"
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "resetMs": { "type": "integer", "description": "Rate limit: ms until limit resets" },
          "retryInMs": { "type": "integer", "description": "Cooldown: ms until next allowed call" },
          "retryInMin": { "type": "integer", "description": "Cooldown: minutes until next allowed call" }
        }
      },
      "WalletAddressRequest": {
        "type": "object",
        "required": ["walletAddress"],
        "properties": {
          "walletAddress": { "$ref": "#/components/schemas/Address" }
        }
      },
...
      "StorySetupRequest": {
        "type": "object",
        "required": ["walletAddress", "name", "vaultUuid"],
        "properties": {
          "walletAddress": { "$ref": "#/components/schemas/Address", "description": "Wallet that will own the IP Asset and the AgentVault agent. Must hold IP for gas (check via /api/wallet/balance)." },
          "name": { "type": "string", "maxLength": 100, "description": "Agent name. Used as the on-chain identifier in AgentVault.createAgent." },
          "vaultUuid": { "type": "integer", "minimum": 1, "description": "CDR vault UUID returned by /api/cdr/store. Passed to AgentVault.createAgent and AgentVault.storeMemory." }
        }
      },
      "StorySetupResponse": {
        "type": "object",
        "required": ["success", "ipId", "licenseTokenId", "readConditionData"],
        "properties": {
          "success": { "type": "boolean" },
          "ipId": { "type": "string", "description": "Address of the registered IP Asset" },
          "licenseTokenId": { "type": "string", "description": "License token ID minted to the user" },
          "licenseTermsId": { "type": "string", "example": "2054" },
          "readConditionData": { "type": "string", "description": "ABI-encoded blob to pass to /api/cdr/store for license-gated recall" },
          "mintTx": { "$ref": "#/components/schemas/TxHash" },
          "registerTx": { "$ref": "#/components/schemas/TxHash" },
          "attachTx": { "$ref": "#/components/schemas/TxHash" },
          "mintLicenseTx": { "$ref": "#/components/schemas/TxHash" },
          "agentId": { "type": "integer", "nullable": true, "description": "On-chain AgentVault agent ID. Set when the caller is the deployer (server-signed). Null for non-deployer users — they sign createAgent client-side and parse the ID from the AgentCreated event." },
          "agentVaultTx": { "$ref": "#/components/schemas/TxHash", "nullable": true, "description": "createAgent tx hash (server-signed path only)" },
          "storeMemoryTx": { "$ref": "#/components/schemas/TxHash", "nullable": true, "description": "storeMemory tx hash (server-signed path only)" },
          "requiresUserSigning": { "type": "boolean", "description": "True if the caller is not the deployer — the client must sign attachLicenseTerms + mintLicenseTokens + createAgent + storeMemory. The licensing pair goes to the IP Account; the AgentVault pair can be batched via EIP-5792 (wallet_sendCalls) if the wallet supports atomic batching." },
          "pendingUserTxs": {
            "type": "object", "nullable": true,
            "description": "Pre-encoded calldata for client-side signing (only when requiresUserSigning=true)",
            "properties": {
              "attachLicenseTerms": { "type": "string", "description": "0x-prefixed calldata for LicensingModule.attachLicenseTerms" },
              "mintLicenseTokens": { "type": "string", "description": "0x-prefixed calldata for LicensingModule.mintLicenseTokens" },
              "createAgent": { "type": "string", "description": "0x-prefixed calldata for AgentVault.createAgent" },
              "storeMemory": { "type": "string", "description": "0x-prefixed calldata for AgentVault.storeMemory" },
              "targets": {
                "type": "object",
                "properties": {
                  "attachLicenseTerms": { "$ref": "#/components/schemas/Address" },
                  "mintLicenseTokens": { "$ref": "#/components/schemas/Address" },
                  "createAgent": { "$ref": "#/components/schemas/Address" },
                  "storeMemory": { "$ref": "#/components/schemas/Address" }
                }
              },
              "licensingModule": { "$ref": "#/components/schemas/Address" },
              "pilTemplate": { "$ref": "#/components/schemas/Address" },
              "licenseTermsId": { "type": "string" },
              "agentVault": { "$ref": "#/components/schemas/Address" },
              "chainId": { "type": "integer", "example": 1315 }
            }
          }
        }
      },
...
      "ContractGetUserAgentsRequest": {
        "type": "object",
        "required": ["action", "userAddress"],
        "properties": {
          "action": { "type": "string", "enum": ["getUserAgents"] },
          "userAddress": { "$ref": "#/components/schemas/Address" }
        }
      },
      "ContractCheckAccessRequest": {
        "type": "object",
        "required": ["action", "agentId", "userAddress"],
        "properties": {
          "action": { "type": "string", "enum": ["checkAccess"] },
          "agentId": { "type": "integer" },
          "userAddress": { "$ref": "#/components/schemas/Address" }
        }
      },
      "ContractGetAgentMemoryCountRequest": {
        "type": "object",
        "required": ["action", "agentId"],
        "properties": {
          "action": { "type": "string", "enum": ["getAgentMemoryCount"] },
          "agentId": { "type": "integer" }
        }
      },
      "ContractResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean" },
          "agents": { "type": "array", "items": { "type": "integer" }, "description": "Returned when action=getUserAgents" },
          "hasAccess": { "type": "boolean", "description": "Returned when action=checkAccess" },
          "count": { "type": "string", "description": "Returned when action=getAgentMemoryCount" }
        }
      },
      "WalletBalanceResponse": {
        "type": "object",
        "required": ["success", "balanceWei", "balanceIp", "hasSufficientFunds"],
        "properties": {
          "success": { "type": "boolean" },
          "address": { "$ref": "#/components/schemas/Address" },
          "balanceWei": { "type": "string", "description": "Balance in wei as a string" },
          "balanceIp": { "type": "string", "description": "Balance in IP (human-readable)" },
          "hasSufficientFunds": { "type": "boolean", "description": "True if balance >= 0.05 IP" },
          "requiredIp": { "type": "string", "example": "0.05" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation failed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "CSRF": {
        "description": "CSRF check failed (Origin/Referer not in allow-list)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "error": "CSRF check failed" } } }
      },
      "RateLimited": {
        "description": "Rate limit or cooldown exceeded",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "ServerError": {
        "description": "Unexpected server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    }
  }
}